+++ plugins/webkul/employees/src/EmployeeServiceProvider.php
@@ -25,9 +25,6 @@ class EmployeeServiceProvider extends PackageServiceProvider
                 '2024_12_11_075011_create_employees_skill_levels_table',
                 '2024_12_11_075017_create_employees_skills_table',
                 '2024_12_11_081046_create_employees_job_positions_table',
-                '2024_12_11_100426_create_employees_calendars_table',
-                '2024_12_11_100435_create_employees_calendar_attendances_table',
-                '2024_12_11_100442_create_employees_calendar_leaves_table',
                 '2024_12_11_120605_create_employees_departure_reasons_table',
                 '2024_12_12_063353_create_employees_employees_table',
                 '2024_12_12_063354_create_employees_employee_skills_table',
