+++ plugins/webkul/employees/src/Filament/Clusters/Configurations/Resources/CalendarResource/Pages/CreateCalendar.php
@@ -2,33 +2,10 @@
 
 namespace Webkul\Employee\Filament\Clusters\Configurations\Resources\CalendarResource\Pages;
 
-use Filament\Notifications\Notification;
-use Filament\Resources\Pages\CreateRecord;
 use Webkul\Employee\Filament\Clusters\Configurations\Resources\CalendarResource;
+use Webkul\Support\Filament\Clusters\Configurations\Resources\CalendarResource\Pages\CreateCalendar as BaseCreateCalendar;
 
-class CreateCalendar extends CreateRecord
+class CreateCalendar extends BaseCreateCalendar
 {
     protected static string $resource = CalendarResource::class;
-
-    public function getSubNavigation(): array
-    {
-        if (filled($cluster = static::getCluster())) {
-            return $this->generateNavigationItems($cluster::getClusteredComponents());
-        }
-
-        return [];
-    }
-
-    protected function getRedirectUrl(): string
-    {
-        return $this->getResource()::getUrl('view', ['record' => $this->getRecord()]);
-    }
-
-    protected function getCreatedNotification(): Notification
-    {
-        return Notification::make()
-            ->success()
-            ->title(__('employees::filament/clusters/configurations/resources/calendar/pages/create-calendar.notification.title'))
-            ->body(__('employees::filament/clusters/configurations/resources/calendar/pages/create-calendar.notification.body'));
-    }
 }
