+++ plugins/webkul/support/database/factories/CalendarAttendanceFactory.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Webkul\Employee\Database\Factories;
+namespace Webkul\Support\Database\Factories;
 
 use Illuminate\Database\Eloquent\Factories\Factory;
-use Webkul\Employee\Models\CalendarAttendance;
 use Webkul\Security\Models\User;
+use Webkul\Support\Models\CalendarAttendance;
 
 class CalendarAttendanceFactory extends Factory
 {
