+++ plugins/webkul/support/resources/lang/en/filament/clusters/configurations/resources/calendar/pages/list-calendar.php
